refactor : 기존 코드 삭제 및 패키지 설정
