+++ src/main/java/bowling/domain/Empty.java
@@ -1,4 +1,4 @@
-package bowling.retry;
+package bowling.domain;
 
 public class Empty implements FrameStatus {
 }
