+++ src/main/java/bowling/domain/FrameManagement.java
@@ -1,4 +1,4 @@
-package bowling.retry;
+package bowling.domain;
 
 public class FrameManagement {
