+++ src/main/java/bowling/domain/FrameStatus.java
@@ -1,4 +1,4 @@
-package bowling.retry;
+package bowling.domain;
 
 public interface FrameStatus {
