+++ src/main/java/bowling/domain/Miss.java
@@ -1,4 +1,4 @@
-package bowling.retry;
+package bowling.domain;
 
 public class Miss implements FrameStatus {
