+++ src/test/java/bowling/domain/FrameTest.java
@@ -1,6 +1,5 @@
-package retry;
+package bowling.domain;
 
-import bowling.retry.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
